Add lb-proxy service
This is the service CRI-O and Argo will point to and use as http proxy.

The service is responsible for routing the traffic to a real http proxy
server, either running in the cluster and/or a external proxy server. If
both a internal and external proxy server are provided, the external
server is used a backup in case the internal is down.