ntp: Ensure the NTP servers aren't competing against each other
This fix a issue where the clients see both of them as "falseticker":
"i.e. its time is inconsistent with a majority of other sources, or
sources specified with the trust option"[1]

This is done with the special "orphan mode", which ensures they aren't
running with the same stratum:
"This option enables a special ‘orphan’ mode, where sources with stratum
equal to the local stratum are assumed to not serve real time. They are
ignored unless no other source is selectable and their reference IDs are
smaller than the local reference ID.

This allows multiple servers in the network to use the same local
configuration and to be synchronised to one another, without confusing
clients that poll more than one server. Each server needs to be
configured to poll all other servers with the local directive. This
ensures only the server with the smallest reference ID has the local
reference active and others are synchronised to it. If that server stops
responding, the server with the second smallest reference ID will take
over when its local reference mode activates (root distance reaches the
threshold configured by the distance option)."[2]

[1] https://chrony.tuxfamily.org/doc/4.1/chronyc.html#sources
[2] https://chrony.tuxfamily.org/doc/4.1/chrony.conf.html#local (orphan)