kafka: Refactor broker config
We will be adding more options to the broker configuration. This
keeps our values file similar to the layout of the kafka broker
configuration making it easier to understand.

With the refactor, we apply defaults to the replication factors,
while keeping the requirement that #replicas <= #brokers intact.
This eliminates some configuration errors in testing senarios
where #brokers is set to 1.